NEW : no-sugar alternative (code only)
Alternative version without any dependency to the
angular-transition-sugar library.

This is the first alternative created.

The current solution (using git branches) should NOT be the final
solution for alternatives management. See #30.

There is still an issue about ng-annotate. So, the current version doesn't
support any uglification, and is therefore not ready for production (or need to use .$inject any time) ! See olov/ng-annotate#248.